Move most frontend handler code to a BundledFileHandler
This is now always compiled as well which greatly reduces the amount of conditionally compiled code and makes compilation faster when working on the handler
