+++ communityvi-server/Cargo.toml
@@ -16,8 +16,7 @@ futures = "0.3"
 gotham = {git = "https://github.com/gotham-rs/gotham", default-features = false, features = ["derive"]}
 governor = {version = "0.3", default-features = false, features = ["std", "jitter"]}
 humantime-serde = "1.0"
-include_dir = {version = "0.6", optional = true, default-features = false}
-lazy_static = "1"
+include_dir = {version = "0.6", default-features = false}
 log = "0.4"
 mime = "0.3"
 mime_guess = "2"
@@ -43,4 +42,4 @@ npm_rs = "0.1"
 
 [features]
 default = []
-bundle-frontend = ["include_dir"]
+bundle-frontend = []
